+++ src/main/java/com/trunksbomb/batteries/inventory/BatteryItemStackHandler.java
@@ -1,9 +1,6 @@
-package com.trunksbomb.batteries.capability;
+package com.trunksbomb.batteries.inventory;
 
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.nbt.ListNBT;
-import net.minecraftforge.common.util.Constants;
 import net.minecraftforge.items.ItemStackHandler;
 
 public class BatteryItemStackHandler extends ItemStackHandler {