+++ src/main/java/com/trunksbomb/batteries/container/BatteryContainerProvider.java
@@ -1,5 +1,6 @@
-package com.trunksbomb.batteries.capability;
+package com.trunksbomb.batteries.container;
 
+import com.trunksbomb.batteries.container.BatteryContainer;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.container.Container;